Decouple types
